[IC-85] refactor: 프로젝트 컴포넌트 패키지 구조 정리 (#54)
