+++ file/src/main/java/com/initcloud/rocket23/infra/redis/config/RedisClientConfig.java
@@ -1,0 +1,33 @@
+package com.initcloud.rocket23.infra.redis.config;
+
+import org.redisson.Redisson;
+import org.redisson.api.RedissonClient;
+import org.redisson.codec.JsonJacksonCodec;
+import org.redisson.config.Config;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class RedisClientConfig {
+
+    @Value("${spring.redis.host}")
+    private String redisHost;
+
+    @Value("${spring.redis.port}")
+    private int redisPort;
+
+    @Value("${spring.redis.password}")
+    private String redisPassword;
+
+    @Bean
+    public RedissonClient redissonClient() {
+        Config config = new Config();
+        config.setCodec(new JsonJacksonCodec())
+                .useSingleServer()
+                .setAddress("redis://" + redisHost + ":" + Integer.toString(redisPort))
+                .setPassword(redisPassword);
+
+        return Redisson.create(config);
+    }
+}
