+++ file/src/main/java/com/initcloud/rocket23/infra/redis/enums/Channels.java
@@ -1,4 +1,4 @@
-package com.initcloud.rocket23.redis.enums;
+package com.initcloud.rocket23.infra.redis.enums;
 
 import lombok.AccessLevel;
 import lombok.Getter;
