+++ file/src/main/java/com/initcloud/rocket23/infra/redis/pubsub/RedisMessageSubscriber.java
@@ -1,14 +1,13 @@
-package com.initcloud.rocket23.redis.pubsub;
-
-import javax.annotation.PostConstruct;
-
-import org.redisson.api.RTopic;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Component;
+package com.initcloud.rocket23.infra.redis.pubsub;
 
 import io.netty.handler.codec.DecoderException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.redisson.api.RTopic;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Component;
+
+import javax.annotation.PostConstruct;
 
 @Slf4j
 @Component
